Moved common logic to common folder.
